fix warning classes.

+++ core/src/commonMain/kotlin/work/socialhub/kmisskey/entity/Role.kt
@@ -11,5 +11,5 @@ class Role {
     lateinit var name: String
     var color: Color? = null
     var iconUrl: String? = null
-    var displayOrder: Long = 0
+    var displayOrder: Int = 0
 }
